Factor out `copy_exact` and `skip_bytes`.
